# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Goodfood 3.0 Users Microservice",
    "version": "1.0.0",
    "title": "Goodfood 3.0 Users Service API"
  },
...
  "paths": {
...
    "/user": {
      "post": {
        "tags": ["user"],
        "summary": "Create user",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid form submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "patch": {
        "tags": ["user"],
        "summary": "Update user record",
        "operationId": "updateUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the user to be updated",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid form submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["user"],
        "summary": "Get user by ID",
        "description": "",
        "operationId": "getUserById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the user to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserComposite"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/{user_id}/preferences": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "user_id",
          "in": "path",
          "required": true,
          "description": "ID of the user"
        },
        {
          "$ref": "#/components/parameters/subscription_id"
        }
      ],
      "get": {
        "summary": "Get the preferences",
        "tags": ["preferences"],
        "operationId": "getPreferences",
        "description": "Gets preferences for a user by default, and for a subscription if the subscription_id is passed",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/PreferenceEntity"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create preferences",
        "tags": ["preferences"],
        "operationId": "createPreferences",
        "description": "Creates preferences for a user by default, and for a subscription if the subscription_id is passed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "description": "An array of the desired preferences"
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/PreferenceEntity"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update preferences",
        "tags": ["preferences"],
        "operationId": "updatePreferences",
        "description": "Updates preferences for a user by default, and for a subscription if the subscription_id is passed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "description": "An array of the desired preferences"
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/PreferenceEntity"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "subscription_id": {
        "in": "query",
        "name": "subscription_id",
        "description": "The subscription ID",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/Subscription/properties/id"
        }
      }
    },
    "schemas": {
...
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string"
          },
          "address_line_1": {
            "type": "string"
          },
          "address_line_2": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "province_code": {
            "type": "string"
          },
          "country_code": {
            "type": "string"
          },
          "postal_code": {
            "type": "string"
          },
          "building_type": {
            "type": "string"
          },
          "special_instructions": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "fsa": {
            "type": "string",
            "example": "K1A"
          }
        },
        "required": ["id"]
      },
...
      "UpdateUserRequestBody": {
        "type": "object",
        "properties": {
          "firebase_id": {
            "type": "string",
            "example": "e1I4xYb9T0U35Od9nfNmaCjgRoS2"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "/^[0-9]{10}$/g"
          },
          "first_name": {
            "type": "string",
            "example": "First"
          },
          "last_name": {
            "type": "string",
            "example": "Last"
          },
          "language": {
            "type": "string",
            "example": "en",
            "enum": ["en", "fr"]
          }
        }
      },
      "CreateUserRequestBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "fsa": {
                "type": "string",
                "pattern": "/^[A-Z][0-9][A-Z]$/g",
                "example": "H3A"
              },
              "referrer_id": {
                "description": "A GF2.0 GiveGet referrer id. 'g' - refers to an email invite. 'l' - refers to a link invite.",
                "type": "string"
              }
            },
            "required": ["fsa", "email", "firebase_id", "language"]
          },
          {
            "$ref": "#/components/schemas/UpdateUserRequestBody"
          }
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firebase_id": {
            "type": "string",
            "example": "e1I4xYb9T0U35Od9nfNmaCjgRoS2",
            "description": "firebase_id comes from the ID generated on Firebase Authentication when the user is registered there. It is a random string of 28 characters with a mix of upper- and lower-case characters and numbers. It must be unique."
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "/^[0-9]{10}$/g"
          },
          "first_name": {
            "type": "string",
            "example": "First"
          },
          "last_name": {
            "type": "string",
            "example": "McLasty"
          },
          "language": {
            "type": "string",
            "example": "en",
            "enum": ["en", "fr"]
          }
        },
        "required": ["id", "firebase_id", "email", "language"]
      },
      "UserComposite": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "properties": {
              "default_postal_code": {
                "type": "string",
                "example": "H4T1K2"
              },
              "addresses": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Address"
                }
              },
              "address_ids": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "deprecated": true,
                "description": "This field is deprecated. Please use the data returned in the `addresses` array."
              },
              "subscriptions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "membership_subscription_ids": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "deprecated": true,
                "description": "This field is deprecated. Please use the data returned in the `subscriptions` array."
              },
              "scheduled_subscription_ids": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "deprecated": true,
                "description": "This field is deprecated. Please use the data returned in the `subscriptions` array."
              },
              "carts": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "integer",
                      "example": "113"
                    },
                    "delivery_date": {
                      "type": "string",
                      "format": "date",
                      "example": "2021-09-24T00:00:00.000Z"
                    },
                    "type": {
                      "type": "string",
                      "enum": ["CURATED", "STANDARD"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["ACTIVE", "CANCELLED", "INCOMPLETE"]
                    },
                    "item_count": {
                      "type": "integer",
                      "example": 3
                    }
                  }
                }
              },
              "preference_tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
      "Membership": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuidv4"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "trial_type": {
            "type": "string",
            "enum": ["DAY", "MONTH", "YEAR"]
          },
          "trial_value": {
            "type": "integer"
          },
          "frequency_type": {
            "type": "string",
            "enum": ["DAY", "MONTH", "YEAR"]
          },
          "frequency_value": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "trial_type",
          "trial_value",
          "frequency_type",
          "frequency_value"
        ]
      },
...
      "Plan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuidv4"
          },
          "number_of_recipes": {
            "type": "integer"
          },
          "number_of_portions": {
            "type": "integer"
          }
        },
        "required": ["id", "number_of_recipes", "number_of_portions"]
      },
...
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuidv4"
          },
          "user_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "card_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "address_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "agreement_id": {
            "type": "integer"
          },
          "state": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CANCELLATION",
              "TRIAL",
              "UNPAID",
              "CANCELLED",
              "PAUSED",
              "SUSPENDED"
            ]
          },
          "subscription_type": {
            "type": "string",
            "enum": ["SCHEDULED", "MEMBERSHIP"]
          },
          "product_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "frequency_type": {
            "type": "string",
            "enum": ["DAY", "MONTH", "YEAR"]
          },
          "frequency_value": {
            "type": "integer"
          },
          "delivery_day": {
            "type": "string",
            "enum": ["MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY"]
          },
          "next_cycle": {
            "type": "string",
            "format": "date"
          },
          "send_notification": {
            "type": "boolean"
          },
          "started_at": {
            "type": "string",
            "format": "date"
          },
          "product": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Membership"
              },
              {
                "$ref": "#/components/schemas/Plan"
              }
            ]
          }
        },
        "required": [
          "id",
          "user_id",
          "card_id",
          "address_id",
          "state",
          "subscription_type",
          "product_id",
          "frequency_type",
          "frequency_value",
          "next_cycle",
          "send_notification",
          "started_at"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "error description"
          },
          "meta": {
            "description": "Structured representation of the error",
            "example": {
              "errors": [["The fsa field is required."]]
            }
          },
          "status_code": {
            "type": "number"
          }
        }
      },
...
      "PreferenceEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "subscription_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "tag": {
            "type": "string"
          }
        },
        "required": ["id", "user_id", "subscription_id", "tag"]
      },
...
    },
...
  }
}